requirements.txt
amqp==2.6.1
asgiref==3.3.4
attrs==21.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==4.4.7
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.3
django-crontab==0.7.1
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
kombu==4.6.11
Markdown==3.3.4
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
mysqlclient==2.0.3
openapi-codec==1.3.2
packaging==20.9
protobuf==3.17.1
pymemcache==3.5.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
structlog==21.1.0
swagger-spec-validator==2.7.3
uritemplate==3.0.1
urllib3==1.26.5
vine==1.3.0
